cast for more pedantic warnings on Linux

+++ src/mbgl/util/geojsonvt.cpp
@@ -497,7 +497,7 @@ void Convert::convertFeature(std::vector<ProjectedFeature> &features, const JSVa
 
         features.push_back(create(tags, projectedType, *geometry));
 
-        printf("features now has %lu items\n", features.size());
+        printf("features now has %lu items\n", (unsigned long)features.size());
     }
 
     else if (type == "MultiPolygon") {
